+++ src/app/(tabs)/profile/index.tsx
@@ -1,5 +1,4 @@
 import { Text, View } from 'react-native';
-import { StyledTextBubble } from '../../../components/styled/TextBubble';
 
 export default function Page() {
   return (
