+++ src/components/styled/ChatBubble/MessageContainer/styles.ts
