+++ src/components/styled/ChatBubble/TextBubbleContainer/styles.ts
